Move filter classes into separate files
