+++ BetaCameras/RealSense2/Filters/Depth2Disparity.cs
@@ -1,0 +1,19 @@
+﻿using Intel.RealSense;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MetriCam2.Cameras.RealSense2Filters
+{
+    public class Depth2Disparity : FilterBase
+    {
+        private DisparityTransform _filter = new DisparityTransform(true);
+
+        protected override VideoFrame ApplyImpl(VideoFrame frame)
+        {
+            return _filter.ApplyFilter(frame);
+        }
+    }
+}
