+++ BetaCameras/RealSense2/Filters/Disparity2Depth.cs
@@ -1,0 +1,19 @@
+﻿using Intel.RealSense;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MetriCam2.Cameras.RealSense2Filters
+{
+    public class Disparity2Depth : FilterBase
+    {
+        private DisparityTransform _filter = new DisparityTransform(false);
+
+        protected override VideoFrame ApplyImpl(VideoFrame frame)
+        {
+            return _filter.ApplyFilter(frame);
+        }
+    }
+}
